Use float32 accumulation in MoE expert weighted sum
All MoE models accumulate the expert weighted sum
`(y * scores[..., None]).sum(axis=-2)` in the input dtype (typically
bfloat16). This causes precision loss during the reduction, leading to
greedy-decoding divergence compared to float32 reference implementations.

Cast gating scores to float32 before the multiply-sum so the entire
reduction runs in float32, then cast the result back to the original
dtype. This matches the pattern already used in `hunyuan.py`.

+++ mlx_lm/models/Klear.py
@@ -146,7 +146,7 @@ class KlearSparseMoeBlock(nn.Module):
             scores = scores / mx.sum(scores, axis=-1, keepdims=True)
         scores = scores.astype(x.dtype)
         expert_out = self.experts(x, inds)
-        y_experts = (expert_out * scores[..., None]).sum(axis=-2)
+        y_experts = (expert_out * scores[..., None].astype(mx.float32)).sum(axis=-2).astype(expert_out.dtype)
         coef = mx.softmax(self.coefficient(x), axis=-1, precise=True)
         shared = self.shared_experts(x)
         y = y_experts * coef[..., :1] + shared * coef[..., 1:]
